requirements.txt
Flask~=2.0.3
requests~=2.27.1
pandas~=1.4.1
Pillow~=9.0.1
docx~=0.2.4